Remove cbc for vs2019.
In tracking down an issue with Anaconda Distribution shortcuts, we found that builds of `pywin32` that used different compilers from the ones used to compile `python` seemed to be the cause. Removing this cbc should use the same version as what is in aggregate. If a python recipe uses another version, we can add it back.